Dark realm changes, ground check fixes, respawning
Changed the Dark Realm to actually lie below the Light Realm.
There is now a void sea that applies drag and pushes objects upwards, as well as damaging the player if they are in it.
The ground check is now a sphere.
The player can now respawn, done by PurgatoryTimer.